+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.0.0] - 2024.08.27
+
 ### Added
 
+- New environment: NetworkGraph
 - Matplotlib render mode for inverted pendulum envs
 
 ## [4.0.1] - 2024.07.30